Add files via upload
We have an advertising dataset of a marketing agency. The goal is to develop a ML algorithm that predicts if a particular user will click on an advertisement. The dataset has 10 features:

Daily Time Spent on Site, Age, Area Income, Daily Internet Usage, Ad Topic Line, City, Male, Country, Timestamp, Clicked on Ad.